Fix attention backends
Current set up is mostly only ideal for Hopper, because it forces either

Flash Attention 2 or Flash Attention 3. Having Flash Attention 3 doesn't
mean you can run it, and Flash Attention 2 is a suboptimal choice for
both Hopper and Blackwell.

`flash_attention` -> `attention` in the A2A op, so we can use SDPA when it's not available.

The new behavior is:
* On Hopper, when users have FAv3 installed, FAv3 should be used
* On Hopper, when users don't have FAv3, the second best choice is cuDNN attention, not FAv2.
* On Blackwell, the best choice is cuDNN attention, not FAv2.
* On Ampere, Ampere RTX, and Ada, FAv2 is the best choice.
* On pre-Ampere, everything else, cuDNN then xFormers are the best choice.
* Also, we don't need to install FAv2, we don't need to ask users to install it. It ships with PyTorch, so we can directly use SDPA instead.