Attention fixes, cuda graphs for V2W
Adds CUDA graphs for Video2World inference, and fixes a few things in
`attention.py`:

There is no use case for varlen, but it was previously using the
varlen API, and the extra ops required to run it broke CUDA graphs,
and added around 3-5% overhead E2E!
We're now removing `q_len` and `k_len` completely.

Also annotating a few more places in attention, and added a dtype check.